pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loadforecast"
version = "0.1.0"
description = "Day-ahead German grid-load forecasting that beats the TSO baseline."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{ name = "Ashish" }]
license = { text = "MIT" }

dependencies = [
  "tensorflow==2.16.2",
  "numpy>=1.26,<2.0",
  "pandas>=2.2,<2.3",
  "pyarrow>=15.0",
  "scikit-learn>=1.4,<1.6",
  "scipy>=1.11",
  "statsmodels>=0.14",
  "holidays>=0.50",
  "requests>=2.32",
  "python-dateutil>=2.9",
  "pytz>=2024.1",
  "pydantic>=2.7",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "streamlit>=1.36",
  "plotly>=5.22",
  "matplotlib>=3.8",
  "mlflow>=2.14",
  "evidently>=0.4.30",
  "tqdm>=4.66",
  "joblib>=1.4",
  "entsoe-py>=0.6",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "ipykernel>=6.29",
  "jupyterlab>=4.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
